feat(app): move policy page into settings
